Adjust preference order
