+++ src/main/res/xml/pref_all.xml
@@ -2,12 +2,21 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
                   xmlns:app="http://schemas.android.com/apk/res/com.github.shadowsocks">
 
+    <PreferenceCategory android:title="@string/global_cat">
+
+        <Preference android:key="profiles" android:summary="@string/profile_summary" android:title="@string/profiles">
+            <intent android:action="com.github.shadowsocks.ProfileManagerActivity"/>
+        </Preference>
+        <SwitchPreference
+                android:key="isAutoConnect"
+                android:summary="@string/auto_connect_summary"
+                android:title="@string/auto_connect"/>
+
+    </PreferenceCategory>
+
     <PreferenceCategory
             android:title="@string/proxy_cat">
 
-        <Preference android:summary="@string/profile_summary" android:title="@string/profiles">
-            <intent android:action="com.github.shadowsocks.ProfileManagerActivity"/>
-        </Preference>
         <com.github.shadowsocks.preferences.SummaryEditTextPreference
                 android:defaultValue="Default"
                 android:key="profileName"
@@ -86,18 +95,16 @@
 
     </PreferenceCategory>
 
-    <PreferenceCategory android:title="@string/global_cat">
+    <PreferenceCategory android:title="@string/misc_cat">
+
         <com.github.shadowsocks.preferences.NatSwitchPreference android:key="isNAT"
                                                                 android:title="@string/nat"
                                                                 android:summary="@string/nat_summary"/>
-        <SwitchPreference
-                android:key="isAutoConnect"
-                android:summary="@string/auto_connect_summary"
-                android:title="@string/auto_connect"/>
         <Preference android:key="recovery" android:title="@string/recovery"
             android:summary="@string/recovery_summary"/>
         <Preference android:key="flush_dnscache" android:title="@string/flush_dnscache"
-            android:summary="@string/flush_dnscache_summary"/>
+                    android:summary="@string/flush_dnscache_summary"/>
+        <Preference android:key="about" android:title="@string/about"/>
 
     </PreferenceCategory>
